+++ nomad/structs/config/tls.go
@@ -141,6 +141,20 @@ func (t *TLSConfig) Copy() *TLSConfig {
 	return new
 }
 
+func (t *TLSConfig) IsEmpty() bool {
+	if t == nil {
+		return true
+	}
+
+	return t.EnableHTTP == false &&
+		t.EnableRPC == false &&
+		t.VerifyServerHostname == false &&
+		t.CAFile == "" &&
+		t.CertFile == "" &&
+		t.KeyFile == "" &&
+		t.VerifyHTTPSClient == false
+}
+
 // Merge is used to merge two TLS configs together
 func (t *TLSConfig) Merge(b *TLSConfig) *TLSConfig {
 	result := t.Copy()
@@ -170,4 +184,22 @@ func (t *TLSConfig) Merge(b *TLSConfig) *TLSConfig {
 		result.RPCUpgradeMode = true
 	}
 	return result
+}
+
+// Equals compares the fields of two TLS configuration objects, returning a
+// boolean indicating if they are the same.
+// It is possible for either the calling TLSConfig to be nil, or the TLSConfig
+// that it is being compared against, so we need to handle both places. See
+// server.go Reload for example.
+func (t *TLSConfig) Equals(newConfig *TLSConfig) bool {
+	if t == nil || newConfig == nil {
+		return t == newConfig
+	}
+
+	return t.EnableRPC == newConfig.EnableRPC &&
+		t.CAFile == newConfig.CAFile &&
+		t.CertFile == newConfig.CertFile &&
+		t.KeyFile == newConfig.KeyFile &&
+		t.RPCUpgradeMode == newConfig.RPCUpgradeMode &&
+		t.VerifyHTTPSClient == newConfig.VerifyHTTPSClient
 }